==> 01_inleiding.Rmd <==
# Inleiding

> Het basisidee van het early-warning meetnet droogte is om in natuurgebieden een vooraf gekozen aantal meetpunten zo goed mogelijk ruimtelijk gebalanceerd te selecteren, rekening houdende met mogelijke verschillen in het gedrag van het grond/oppervlaktewaterpeil t.a.v. een neerslagtekort. Voor deze meetpunten kan op basis van actuele informatie een indicator (cfr. droogte-indicator) worden berekend.

Het project omhelst : 

* uitwerking van een methodiek die een ruimtelijk gebalanceerde selectie van actuele en potentiële meetlocaties mogelijk maakt 
* bepalen van referentiewaarden voor de actuele meetlocaties

Met natuurgebieden worden hier alle locaties met grondwaterafhankelijke habitattypes en rbb's bedoeld, gelegen binnen of buiten het Natura 2000 netwerk. In overeenstemming met het meetnet natuurlijk milieu (MNM) voor de milieudruk 'verdroging via grondwater' worden volgende habitattypen en rbb's als grondwaterafhankelijk beschouwd (tabel \@ref(tab:gw-typen)). Hierna worden deze typen afgekort tot gaHT. 

```{r gw-typen, eval=TRUE, include = TRUE, echo=FALSE, warning=FALSE}
kobj <- kable(gw_types %>% 
        dplyr::select(type, groupnr) %>% 
        inner_join(read_namelist(lang = "nl") %>% 
                     dplyr::select(code, shortname),
                   by = c("type" = "code")) %>% 
        arrange(type) %>% 
        dplyr::select(1,3,2) %>% 
        rename("habitat(sub)type of rbb" = type,
               "(verkorte) naam" = shortname,
               "GT-groep: nummer" = groupnr
                          ),

      caption =  "grondwaterafhankelijke habitattypen en rbb's") 
if (pdfmaken == TRUE) {
  kobj
} else {
  kobj   %>%
  kable_styling(bootstrap_options = c("striped", "hover", "condensed", "responsive"),
                full_width = F,
                position = "left",
                font_size = 12,
                fixed_thead = T) %>%
  #row_spec(0, angle = -45)  %>% #fixeer veldnamen
  scroll_box(height = "200px")
}
```
  

Een min of meer ruimtelijk gebalanceerde set van bestaande meetlocaties maakt het mogelijk om conclusies te trekken die benaderend representatief zijn voor heel het studiegebied.
Via deze benadering wordt ook een goede synergie bekomen met de implementatie van de zg. 'overgangsfase' van het grondwatermeetnet in MNM [@vanderhaeghe_meetnetten_2018: hoofdstuk 6].

Er wordt met dit meetnet niet beoogd om specifiek voor elk gaHT te bepalen of het aan droogte onderhevig is. Het meetnet wil een globale uitspraak faciliteren.
Hierbij achten we het belangrijk om rekening te houden met de mogelijke verschillen in responstijd ('hoe snel laat een neerslagtekort zich voelen in een wijziging van het grondwaterpeil') op droogte. De responstijd is onder meer afhankelijk van de landschappelijke positie. Vegetaties die voorkomen in lokaal-grondwatergevoede systemen zullen sneller een droogte gewaar worden dan vegetaties die gebonden zijn aan (regionale) systemen met een permanente aanvoer van grondwater.
We veronderstellen hierbij dat er een relatie bestaat tussen het hydrologisch regime en de responstijd: permanent grondwatergevoede systemen reageren trager dan systemen met een tijdelijke of zwakke grondwatervoeding. 

We vertrekken hiervoor van de 5-delige indeling in grondwatertypegroepen (GT-groep) die toegepast wordt in het MNM (tabel \@ref(tab:GTgroepen)). In de vorige tabel \@ref(tab:gw-typen)) is weergegeven tot welke GT-groep een gaHT gerekend wordt.
```{r GTgroepen}
gw_types_groups <- gw_types %>%
  distinct(groupnr, typegroup_name) %>% 
  inner_join(data.frame(groupnr = seq(1:5), gewenst_aantal_locaties = minaantal_tub_group), by = "groupnr") %>% 
  arrange(groupnr) %>% 
  rename("GT-groep: nummer" = groupnr,
         "GT-groep: naam" = typegroup_name,
         "Gewenst aantal meetlocaties" = gewenst_aantal_locaties)

kobj <- kable(gw_types_groups %>%  select(1:2), 
      caption = "Grondwatertype-groepen") 

if (pdfmaken == TRUE) {
  kobj
} else {
  kobj   %>% 
    kable_styling(bootstrap_options = c("striped", "hover", "condensed", "responsive"), 
                  full_width = TRUE, 
                  position = "left",
                  font_size = 12,
                  fixed_thead = T) 
    #row_spec(0, angle = -45)  %>% #fixeer veldnamen
     # scroll_box(height = "200px") 
}
```
    

De GT-groepen lenen zich heel goed om een stratificatie uit te voeren op basis van de verwachte responssnelheid van een locatie op droogte. 
De permanent gevoede locaties (GT-groep 1) zullen naar verwachting trager reageren dan de tijdelijk of zwak gevoede locaties. 
GT-groep 5 bundelt grondwaterafhankelijke HT die zodanig breed gedefinieerd zijn dat ze op sommige locaties grondwatergevoed zijn, maar soms tot meestal niet.
Het voordeel van deze groep is dat ze vermoedelijk het snelst op droogte zal reageren.
Ze houdt echter ook nadelen in, nl. dat je bij het uitzetten van een nieuw punt niet op voorhand weet of het wel grondwaterafhankelijk is en dat sommige typen heel variabel zijn m.b.t. het grondwaterregime.
We oordelen dat deze nadelen groter zijn dan het mogelijke voordeel van een snelle signalering.
De GT-groep 5 wordt daarom niet meegenomen in de verdere analyse.

Door een stratificatie over de GT-groepen toe te passen kan een balans worden verzekerd. 
Een stratificatie laat toe om een globale uitspraak te doen.
We raden echter ook aan in de beoordeling elke GT-groep apart mee te nemen, omdat ze verschillende signalen zullen geven. 

```{r habitat selection, include=FALSE}

## Verspreiding van de verdrogingsgevoelige typen in Vlaanderen volgens de habitatkaart.
# if (file.exists(file.path(".","data","local", "habmap_terr_gw.gpkg")) == FALSE | params$refresh_data == 2) {
#     drive_download(drive_get(id = "1nxnpfE3Eh4eCiM2VinGYMJE55qD4Az1c"), 
#                    path = file.path(".","data","local", "habmap_terr_gw.gpkg"), overwrite = TRUE)
# }

habmap_polygons_gw <- habmap_polygons_gw %>%
  mutate(polygon_id = factor(.data$polygon_id),
         type = factor(.data$type),
         typegroup_name = factor(.data$typegroup_name)
         )
```
